Update and rename task02.cpp to task02-03.cpp
